remove now-dead bootstrap code: HardwareMonitor, Agent, Config
