+++ sled-agent/src/bootstrap/http_entrypoints.rs
@@ -7,10 +7,9 @@
 //! Note that the bootstrap agent also communicates over Sprockets,
 //! and has a separate interface for establishing the trust quorum.
 
-use super::agent::BootstrapError;
 use super::rack_ops::RssAccess;
+use super::BootstrapError;
 use super::RssAccessError;
-use crate::bootstrap::agent::Agent;
 use crate::bootstrap::params::RackInitializeRequest;
 use crate::bootstrap::rack_ops::{RackInitId, RackResetId};
 use crate::storage_manager::StorageResources;
